+++ spinn_utilities/citation/__init__.py
@@ -1,5 +1,5 @@
 from .citation_updater_and_doi_generator import CitationUpdaterAndDoiGenerator
-from .tool_citation_generation import CitationAggregator
+from .citation_aggregator import CitationAggregator
 from .generate_aggregate import generate_aggregate
 
 __all__ = ["CitationUpdaterAndDoiGenerator", "CitationAggregator",
